v1.2 - Changed shortcodes to [audio] and [video]. Separate button for inserting media player. And huge bunch of backend fixes and changes.

+++ wp-nonverblaster-hover.php
@@ -5,7 +5,7 @@ PluginURI: https://github.com/jacobbuck/wp-nonverblaster-hover
 Description: Play video and audio files using the NonverBlaster:hover flash player, or HTML5 fallback for mobile.
 Author: Jacob Buck
 Author URI: http://jacobbuck.co.nz/
-Version: 1.1.7
+Version: 1.2
 */
 
 class WPNonverBlasterHover {
@@ -20,10 +20,12 @@ class WPNonverBlasterHover {
 		"video_height" => "",
 		"video_crop" => false,
 		"video_default_hd" => false	);
-	private $version = "1.1.7";
+	private $fallback_width;
+	private $fallback_height;
+	private $version = "1.2";
 	
 	public function __construct () {
-		$this->options = json_decode(get_option("wpnbh_options"));
+		$this->activate();
 		// Actions
 		add_action("init", array($this, "init"));
 		add_action("wp_enqueue_scripts", array($this, "wp_enqueue_scripts"));
@@ -31,32 +33,39 @@ class WPNonverBlasterHover {
 		add_action("admin_enqueue_scripts", array($this, "admin_enqueue_scripts"));
 		add_action("admin_menu", array($this, "admin_menu"));
 		// Filters
-		add_filter("audio_send_to_editor_url", array($this, "insert_audioplayer_shortcode"), 20, 3);
-		add_filter("video_send_to_editor_url", array($this, "insert_videoplayer_shortcode"), 20, 3);
-		add_filter("media_send_to_editor", array($this, "insert_mediaplayer_shortcode"), 20, 3);
+		add_filter("attachment_fields_to_edit", array($this, "insert_shortcode_button_field"), 10, 2);
+		add_filter("attachment_fields_to_save", array($this, "send_shortcode_to_editor"), 10, 2);
 		add_filter("plugin_action_links", array($this, "add_settings_link"), 10, 2);
 		// Shortcodes
-		add_shortcode("audioplayer", array($this, "audioplayer_shortcode_func"));
-		add_shortcode("videoplayer", array($this, "videoplayer_shortcode_func"));
+		add_shortcode("audio", array($this, "audio_shortcode_func"));
+		add_shortcode("video", array($this, "video_shortcode_func"));
+		// Shortcodes (legacy)
+		add_shortcode("audioplayer", array($this, "audio_shortcode_func"));
+		add_shortcode("videoplayer", array($this, "video_shortcode_func"));
 	}
 	
 	/* Install */
 	
-	public function install () {
-		add_option("wpnbh_options", json_encode($this->default_options));
+	public function activate () {
+		$this->options = get_option("wpnbh_options");
+		if (empty($this->options) || ! is_array($this->options)) {
+			// Get default options
+			$this->options = $this->default_options;
+			// Set defualt options
+			delete_option("wpnbh_options");
+			add_option("wpnbh_options", $this->options);
+		}
 	}
-	
-	/* Uninstall */
-	
-	public function uninstall () {
-		delete_option("wpnbh_options");
-	}
-	
+		
 	/* Plugin Scripts & Styles */
 		
 	public function init () {
 		// Front end
 		wp_register_script("wp-nonverblaster-hover", plugins_url("/assets/wp-nonverblaster-hover.js", __FILE__), array("swfobject"), $this->version);
+		// Get default sizes
+		$wp_embed_defaults = wp_embed_defaults();
+		$this->fallback_width = empty($this->options["video_width"]) ? $wp_embed_defaults["width"] : $this->options["video_width"];
+		$this->fallback_height = empty($this->options["video_height"]) ? round($this->fallback_width * .5625) : $this->options["video_height"];
 	}
 	
 	public function wp_enqueue_scripts () {
@@ -67,35 +76,49 @@ class WPNonverBlasterHover {
 		));
 	}
 	
-	/* Media Library inserts Shortcods */
+	/* Insert shortcode in media library */
 	
-	public function insert_audioplayer_shortcode ($html, $url, $post_title) {
-		return "[audioplayer src=\"$url\" title=\"".htmlspecialchars($post_title)."\"]";
-	}
-	
-	public function insert_videoplayer_shortcode ($html, $url, $post_title) {
-		return "[videoplayer src=\"$url\" title=\"".htmlspecialchars($post_title)."\"]";
-	}
-	
-	public function insert_mediaplayer_shortcode ($html, $send_id, $attachment) {
-		$mime_type = get_post_mime_type($send_id);
-		$url = wp_get_attachment_url($send_id);
-		$post_title = htmlspecialchars($attachment["post_title"]);
-		switch (substr($mime_type, 0, strpos($mime_type, "/"))) {
-			case "audio":
-				return "[audioplayer src=\"$url\" title=\"$post_title\"]";
+	public function insert_shortcode_button_field ($form_fields, $attachment) {
+		switch (reset(explode("/", $attachment->post_mime_type))) {
+			case "audio" :
+				$form_fields["wpnbh"] = array(
+					"label" => "",
+					"input" => "html",
+					"html" => get_submit_button(__("Insert Audio Player into Post"), "button", "wpnbhsend[".$attachment->ID."]", false )
+				);
 				break;
-			case "video":
-				return "[videoplayer src=\"$url\" title=\"$post_title\"]";
+			case "video" :
+				$form_fields["wpnbh"] = array(
+					"label" => "",
+					"input" => "html",
+					"html" => get_submit_button(__("Insert Video Player into Post"), "button", "wpnbhsend[".$attachment->ID."]", false )
+				);
 				break;
-			default:
-				return $html;
 		}
+		return $form_fields;
+	}
+	
+	public function send_shortcode_to_editor ($post, $attachment) {
+		if (isset($_POST["wpnbhsend"])) {
+			$src = wp_get_attachment_url($post["ID"]);
+			$title = htmlspecialchars($attachment["post_title"]);
+			switch (reset(explode("/", $post["post_mime_type"]))) {
+				case "audio":
+					media_send_to_editor("[audio src=\"$src\" title=\"$title\"]");
+					break;
+				case "video":
+					media_send_to_editor("[video src=\"$src\" title=\"$title\"]");
+					break;
+				default:
+					return $post;
+			}
+		}
+		return $post;
 	}
 	
 	/* Shortcode Functions */
 	
-	public function audioplayer_shortcode_func ($atts) {
+	public function audio_shortcode_func ($atts) {
 		extract(shortcode_atts(array(
 			"src" => "",
 			"title" => "",
@@ -103,16 +126,15 @@ class WPNonverBlasterHover {
 			"controls" => "controls",
 			"loop" => "",
 		), $atts));
-		$width = $this->options->audio_width . (substr($this->options->audio_width, -1) == "%" ? "" : "px");
+		$width = $this->options["audio_width"] . (substr($this->options["audio_width"], -1) == "%" ? "" : "px");
 		return "<audio src=\"$src\"" .
 			($autoplay ? " autoplay=\"autoplay\" " : "") .
 			($controls ? " controls=\"controls\" " : "") .
 			($loop ? " loop=\"loop\" " : "") .
-			" class=\"nonverblaster nonverblaster-audio\" style=\"width:$width\" title=\"$title\" id=\"nonverblaster_".md5(time().$src)."\"><a href=\"http://get.adobe.com/flashplayer\" target=\"_blank\">Adobe Flash Player</a> is required to listen to audio.</audio>";
+			" class=\"nonverblaster nonverblaster-audio\" preload=\"none\" style=\"width:$width\" title=\"$title\" id=\"nonverblaster_".md5(time().$src)."\"><a href=\"http://get.adobe.com/flashplayer\" target=\"_blank\">Adobe Flash Player</a> is required to listen to audio.</audio>";
 	}
 	
-	public function videoplayer_shortcode_func ($atts) {
-		$wp_embed_defaults = wp_embed_defaults();
+	public function video_shortcode_func ($atts) {
 		extract(shortcode_atts(array(
 			"src" => "",
 			"hdsrc" => "",
@@ -121,8 +143,8 @@ class WPNonverBlasterHover {
 			"autoplay" => "",
 			"controls" => "controls",
 			"loop" => "",
-			"width" => $this->options->video_width ? $this->options->video_width : $wp_embed_defaults["width"],
-			"height" => $this->options->video_height ? $this->options->video_height : round($wp_embed_defaults["width"] * .5625) /* 16:9 */
+			"width" => $this->fallback_width,
+			"height" => $this->fallback_height
 		), $atts));
 		return "<video src=\"$src\"" .
 			($hdsrc ? " data-hdsrc=\"$hdsrc\" " : "") .
@@ -130,36 +152,34 @@ class WPNonverBlasterHover {
 			($controls ? " controls=\"controls\" " : "") .
 			($loop ? " loop=\"loop\" " : "") .
 			($poster ? " poster=\"$poster\" " : "") .
-			" width=\"$width\" height=\"$height\" class=\"nonverblaster nonverblaster-video\" title=\"$title\" id=\"nonverblaster_".md5(time().$src)."\"><a href=\"http://get.adobe.com/flashplayer\" target=\"_blank\">Adobe Flash Player</a> is required to watch video.</video>";
+			" width=\"$width\" height=\"$height\" preload=\"none\" class=\"nonverblaster nonverblaster-video\" title=\"$title\" id=\"nonverblaster_".md5(microtime().$src)."\"><a href=\"http://get.adobe.com/flashplayer\" target=\"_blank\">Adobe Flash Player</a> is required to watch video.</video>";
 	}
 	
 	/* Plugin Options Page */
 	
 	public function admin_init () {
-		if (! wp_verify_nonce($_POST["wpnbh_nonce"], plugin_basename( __FILE__ ))) 
-			return false;
+		if (empty($_POST["wpnbh_nonce"]) || ! wp_verify_nonce($_POST["wpnbh_nonce"], plugin_basename( __FILE__ ))) 
+			return;
 		$posted = $_POST["wpnbh"];
 		/* Filter Hex Colors */
-		foreach (array("player_back_color", "control_color", "control_back_color") as $name) {
-			$posted[$name] = preg_replace('/(^#[a-fA-F0-9]{6})/', '$1', $posted[$name]);
-			$posted[$name] = $posted[$name] ? strtolower($posted[$name]) : $this->default_options[$name];
-		}
+		foreach (array("player_back_color", "control_color", "control_back_color") as $name)
+			$posted[$name] = strtolower(preg_replace('/(^#[a-fA-F0-9]{6})/', '$1', $posted[$name]));
 		/* Filter Sizes */
-		foreach (array("audio_width", "video_width", "video_height") as $name) {
-			$posted[$name] = preg_replace('/(^[0-9]+%?)/', '$1', $posted[$name]);
-			$posted[$name] = $posted[$name] ? strtolower($posted[$name]) : $this->default_options[$name];
-		}
+		foreach (array("audio_width", "video_width", "video_height") as $name)
+			$posted[$name] = strtolower(preg_replace('/(^[0-9]+%?)/', '$1', $posted[$name]));
 		/* Filter Checkboxes */
-		foreach (array("video_crop", "video_default_hd") as $name) {
-			$posted[$name] = !! $posted[$name];
-		}
+		foreach (array("video_crop", "video_default_hd") as $name)
+			$posted[$name] = ! empty($posted[$name]);
+		/* Revert Empty Options To Default */
+		foreach (array("player_back_color", "control_color", "control_back_color", "audio_width", "video_width", "video_height") as $name)
+			$posted[$name] = empty($posted[$name]) ? $posted[$name] : $this->default_options[$name];
 		/* Update Option */
-		update_option("wpnbh_options", json_encode($posted));
-		wp_redirect(admin_url("/options-general.php?page=options-wpnbh&settings-updated=true"));
+		update_option("wpnbh_options", $posted);
+		wp_redirect(admin_url("options-general.php?page=wpnbh&settings-updated=true"));
 	}
 	
 	public function admin_menu() {
-		add_options_page("NonvernBlaster:hover Settings", "NonvernBlaster:hover", "manage_options", "options-wpnbh", array($this, "plugin_options"));
+		add_options_page("NonvernBlaster:hover Settings", "NonvernBlaster:hover", "manage_options", "wpnbh", array($this, "plugin_options"));
 	}
 	
 	public function admin_enqueue_scripts ($hook) {
@@ -171,7 +191,7 @@ class WPNonverBlasterHover {
 	
 	public function add_settings_link ($links, $file) {
 		if (strstr(__FILE__, $file)) 
-			$links[] = "<a href=\"options-general.php?page=options-wpnbh\">".__("Settings")."</a>";
+			$links[] = "<a href=\"".admin_url("options-general.php?page=wpnbh")."\">".__("Settings")."</a>";
 		return $links;
 	}
 	
@@ -185,12 +205,7 @@ class WPNonverBlasterHover {
 		<div class="wrap options-wpnbh">
 			<div id="icon-options-general" class="icon32"><br></div>
 			<h2>NonvernBlaster<span style="font-style:italic;font-weight:lighter">:hover</span> Settings</h2>
-			
-			<?php if (isset($_GET["settings-updated"]) && $_GET["settings-updated"] === "true") : ?>
-			<div class="updated settings-error" id="setting-error-settings_updated"> 
-			<p><strong>Settings saved.</strong></p></div>
-			<?php endif; ?>
-			
+						
 			<form action="" method="post">
 				<input type="hidden" name="option_page" value="wpnbh">
 				<input type="hidden" name="action" value="update">
@@ -201,21 +216,21 @@ class WPNonverBlasterHover {
 						<tr valign="top">
 							<th scope="row">Background color</th>
 							<td><fieldset><legend class="screen-reader-text"><span>Default audio size</span></legend>
-							<input name="wpnbh[player_back_color]" type="text" id="wpnbh_player_back_color" value="<?php echo $this->options->player_back_color; ?>" class="small-text color">
+							<input name="wpnbh[player_back_color]" type="text" id="wpnbh_player_back_color" value="<?php echo $this->options["player_back_color"]; ?>" class="small-text color">
 							<div class="colorpicker" id="wpnbh_player_back_color-farbtastic"></div>
 							</fieldset></td>
 						</tr>
 						<tr valign="top">
 							<th scope="row">Control color</th>
 							<td><fieldset><legend class="screen-reader-text"><span>Default audio size</span></legend>
-							<input name="wpnbh[control_color]" type="text" id="wpnbh_control_color" value="<?php echo $this->options->control_color; ?>" class="small-text color">
+							<input name="wpnbh[control_color]" type="text" id="wpnbh_control_color" value="<?php echo $this->options["control_color"]; ?>" class="small-text color">
 							<div class="colorpicker" id="wpnbh_control_color-farbtastic"></div>
 							</fieldset></td>
 						</tr>
 						<tr valign="top">
 							<th scope="row">Control back colour</th>
 							<td><fieldset><legend class="screen-reader-text"><span>Default audio size</span></legend>
-							<input name="wpnbh[control_back_color]" type="text" id="wpnbh_control_back_color" value="<?php echo $this->options->control_back_color; ?>" class="small-text color">
+							<input name="wpnbh[control_back_color]" type="text" id="wpnbh_control_back_color" value="<?php echo $this->options["control_back_color"]; ?>" class="small-text color">
 							<div class="colorpicker" id="wpnbh_control_back_color-farbtastic"></div>
 							</fieldset></td>
 						</tr>
@@ -229,16 +244,16 @@ class WPNonverBlasterHover {
 							<th scope="row">Audio player size</th>
 							<td><fieldset><legend class="screen-reader-text"><span>Audio player size</span></legend>
 							<label for="wpnbh_audio_width">Width</label>
-							<input name="wpnbh[audio_width]" type="text" id="wpnbh_audio_width" value="<?php echo $this->options->audio_width; ?>" class="small-text">
+							<input name="wpnbh[audio_width]" type="text" id="wpnbh_audio_width" value="<?php echo $this->options["audio_width"]; ?>" class="small-text">
 							</fieldset></td>
 						</tr>
 						<tr valign="top">
 							<th scope="row">Video player default size</th>
 							<td><fieldset><legend class="screen-reader-text"><span>Video player default size</span></legend>
 							<label for="wpnbh_video_width">Width</label>
-							<input name="wpnbh[video_width]" type="text" id="wpnbh_video_width" value="<?php echo $this->options->video_width; ?>" class="small-text">
+							<input name="wpnbh[video_width]" type="text" id="wpnbh_video_width" value="<?php echo $this->options["video_width"]; ?>" class="small-text">
 							<label for="wpnbh_video_height">Height</label>
-							<input name="wpnbh[video_height]" type="text" id="wpnbh_video_height" value="<?php echo $this->options->video_height; ?>" class="small-text">
+							<input name="wpnbh[video_height]" type="text" id="wpnbh_video_height" value="<?php echo $this->options["video_height"]; ?>" class="small-text">
 							</fieldset>
 							If the width or height value is left blank, the default maximum embed size will be used.
 							</td>
@@ -250,10 +265,10 @@ class WPNonverBlasterHover {
 				<table class="form-table">
 					<tbody>
 						<tr valign="top">
-							<th scope="row" colspan="2" class="th-full"><label for="wpnbh_video_crop"><input name="wpnbh[video_crop]" type="checkbox" id="wpnbh_video_crop" value="true" <?php echo $this->options->video_crop ? " checked=\"checked\" " : ""; ?>> Crop video to fit player.</label></th>
+							<th scope="row" colspan="2" class="th-full"><label for="wpnbh_video_crop"><input name="wpnbh[video_crop]" type="checkbox" id="wpnbh_video_crop" value="true" <?php echo $this->options["video_crop"] ? " checked=\"checked\" " : ""; ?>> Crop video to fit player.</label></th>
 						</tr>
 						<tr valign="top">
-							<th scope="row" colspan="2" class="th-full"><label for="wpnbh_video_default_hd"><input name="wpnbh[video_default_hd]" type="checkbox" id="wpnbh_video_default_hd" value="true" <?php echo $this->options->video_default_hd ? " checked=\"checked\" " : ""; ?>> Enable HD on default, when HD video is available.</label></th>
+							<th scope="row" colspan="2" class="th-full"><label for="wpnbh_video_default_hd"><input name="wpnbh[video_default_hd]" type="checkbox" id="wpnbh_video_default_hd" value="true" <?php echo $this->options["video_default_hd"] ? " checked=\"checked\" " : ""; ?>> Enable HD on default, when HD video is available.</label></th>
 						</tr>
 					</tbody>
 				</table>
@@ -285,6 +300,3 @@ class WPNonverBlasterHover {
 }
 
 $wpnbh = new WPNonverBlasterHover;
-
-register_activation_hook(__FILE__, array($wpnbh, "install"));
-register_deactivation_hook(__FILE__, array($wpnbh, "uninstall"));
